requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
numpy>=1.24.3
plotly>=5.14.1
pydantic>=2.0.0
pyyaml>=6.0
openpyxl>=3.1.2
pytest>=7.3.1
